complete transfer flow

+++ transfer/src/main/java/com/pmb/transfer/presentation/transfer_amount/viewmodel/TransferAmountViewEvents.kt
@@ -3,5 +3,5 @@ package com.pmb.transfer.presentation.transfer_amount.viewmodel
 import com.pmb.core.platform.BaseViewEvent
 
 interface TransferAmountViewEvents: BaseViewEvent {
-    data class NavigateToDestinationType(val amount: Long) : TransferAmountViewEvents
+    data class NavigateToDestinationType(val amount: Double) : TransferAmountViewEvents
 }
