+++ transfer/src/main/java/com/pmb/transfer/presentation/transfer_confirm/viewmodel/TransferConfirmViewActions.kt
@@ -14,7 +14,7 @@ sealed interface TransferConfirmViewActions : BaseViewAction {
     data class SelectCardBank(val item: CardBankEntity) : TransferConfirmViewActions
     data class UpdateData(
         val account: TransactionClientBankEntity?,
-        val amount: Long?,
+        val amount: Double?,
         val transferMethod: TransferMethodEntity?,
         val reason: ReasonEntity?
     ) : TransferConfirmViewActions
